chore: 보드 패키지에 있던 코멘트 관련 파일 이동
CommenteResponseDto 변경으로 관련 코멘트 수정 필요
