+++ src/main/java/com/example/newsfeed/dto/boardDto/response/BoardResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.newsfeed.dto.boardDto.response;
 
-import com.example.newsfeed.dto.boardDto.CommentResponseDto;
+import com.example.newsfeed.dto.comment.responseDto.CommentResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.data.domain.Page;
