+++ src/main/java/com/example/newsfeed/dto/comment/responseDto/CommentResponseDto.java
@@ -13,5 +13,7 @@ public class CommentResponseDto {
     private String img_url;
     private String name;
     private String contents;
+    private int likeCnt;
+    private boolean like;
     private LocalDateTime updatedAt;
 }
