--- src/main/java/com/example/newsfeed/entity/board/Comment.java
@@ -1,29 +1,0 @@
-package com.example.newsfeed.entity.board;
-
-import com.example.newsfeed.common.entity.BaseEntity;
-import com.example.newsfeed.entity.user.User;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Table(name = "comment")
-@NoArgsConstructor
-public class Comment extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String contents;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "board_id")
-    private Board board;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-}
