--- src/main/java/com/example/newsfeed/service/board/CommentService.java
@@ -1,6 +1,0 @@
-package com.example.newsfeed.service.board;
-
-
-public interface CommentService {
-    void likes(Long commentId, Long userId);
-}
